+++ test/python/test_ibmqjob.py
@@ -9,27 +9,33 @@
 
 """IBMQJob Test."""
 
-import unittest
 import time
+import unittest
 from concurrent import futures
+
 import numpy
 from scipy.stats import chi2_contingency
 
 from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
 from qiskit import transpiler
-from qiskit.backends.ibmq import IBMQProvider
-from qiskit.backends.ibmq.ibmqjob import IBMQJob, IBMQJobError
-from qiskit.backends.ibmq.ibmqbackend import IBMQBackendError
 from qiskit.backends import JobStatus
+from qiskit.backends.ibmq import IBMQProvider
+from qiskit.backends.ibmq.ibmqbackend import IBMQBackendError
+from qiskit.backends.ibmq.ibmqjob import IBMQJob, IBMQJobError
 from .common import requires_qe_access, QiskitTestCase, slow_test
 
 
-def lowest_pending_jobs(backends):
-    """Returns the backend with lowest pending jobs."""
-    backends = filter(lambda x: x.status.get('available', False), backends)
-    by_pending_jobs = sorted(backends,
-                             key=lambda x: x.status['pending_jobs'])
-    return by_pending_jobs[0]
+def _least_busy(backends):
+    """Helper version of `wrapper.least_busy()` that works on instances.
+    Args:
+        backends (list[BaseBackend]): list of backends.
+
+    Returns:
+        BaseBackend: least busy backend instance.
+    """
+    return min([b for b in backends if
+                b.status['available'] and 'pending_jobs' in b.status],
+               key=lambda b: b.status['pending_jobs'])
 
 
 class TestIBMQJob(QiskitTestCase):
@@ -89,9 +95,10 @@ class TestIBMQJob(QiskitTestCase):
 
     @slow_test
     def test_run_device(self):
-        backends = self._provider.available_backends({'simulator': False})
+        backends = [backend for backend in self._provider.available_backends()
+                    if not backend.configuration['simulator']]
         self.log.info('devices: %s', [b.name for b in backends])
-        backend = lowest_pending_jobs(backends)
+        backend = _least_busy(backends)
         self.log.info('using backend: %s', backend.name)
         qobj = transpiler.compile(self._qc, backend)
         shots = qobj['config']['shots']
@@ -164,8 +171,9 @@ class TestIBMQJob(QiskitTestCase):
 
     @slow_test
     def test_run_async_device(self):
-        backends = self._provider.available_backends({'simulator': False})
-        backend = lowest_pending_jobs(backends)
+        backends = [backend for backend in self._provider.available_backends()
+                    if not backend.configuration['simulator']]
+        backend = _least_busy(backends)
         self.log.info('submitting to backend %s', backend.name)
         num_qubits = 5
         qr = QuantumRegister(num_qubits, 'qr')
@@ -210,7 +218,8 @@ class TestIBMQJob(QiskitTestCase):
     def test_cancel(self):
         if not self._using_hub:
             self.skipTest('job cancellation currently only available on hubs')
-        backends = self._provider.available_backends({'simulator': False})
+        backends = [backend for backend in self._provider.available_backends()
+                    if not backend.configuration['simulator']]
         self.log.info('devices: %s', [b.name for b in backends])
         backend = backends[0]
         self.log.info('using backend: %s', backend.name)
@@ -258,8 +267,7 @@ class TestIBMQJob(QiskitTestCase):
         self.assertTrue(job.backend_name == backend_name)
 
     def test_get_jobs_from_backend(self):
-        backends = self._provider.available_backends()
-        backend = lowest_pending_jobs(backends)
+        backend = _least_busy(self._provider.available_backends())
         start_time = time.time()
         job_list = backend.jobs(limit=5, skip=0)
         self.log.info('time to get jobs: %0.3f s', time.time() - start_time)
@@ -278,8 +286,9 @@ class TestIBMQJob(QiskitTestCase):
         self.assertTrue(job.result().get_counts() == rjob.result().get_counts())
 
     def test_retrieve_job_error(self):
-        backends = self._provider.available_backends({'simulator': False})
-        backend = lowest_pending_jobs(backends)
+        backends = [backend for backend in self._provider.available_backends()
+                    if not backend.configuration['simulator']]
+        backend = _least_busy(backends)
         self.assertRaises(IBMQBackendError, backend.retrieve_job, 'BAD_JOB_ID')