+++ CHANGELOG.rst
@@ -22,10 +22,13 @@ Added
 -----
 - Add German introductory documentation (``doc/de``) (#592).
 - Add ``unregister()`` for removing previously registered providers (#584).
+- Adding backend filtering by least busy (#575).
 
 Changed
 -------
+- Remove backend filtering in individual providers, keep only in wrapper (#575).
 - Single source of version information (#581)
+- Bumped IBMQuantumExperience dependency to 1.9.6 (#600).
 
 Removed
 -------